+++ testlib/src/main/resources/license/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Test package-info.java file.
+ * <p>
+ * Used only by integration tests.
+ *
+ * @since 1.0
+ */
+package com.diffplug.spotless;